# it_drinks_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🍸 IT Drinks API",
        "description": "API RESTful para la gestión de cócteles e ingredientes con Passport",
        "contact": {
            "name": "Rafa",
            "email": "rafa@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/cocktails": {
            "get": {
                "tags": [
                    "Cocktails"
                ],
                "summary": "Listar todos los cócteles",
                "description": "Devuelve una lista de todos los cócteles con sus ingredientes. Accesible para cualquier usuario autenticado.",
                "operationId": "4998c94931df0edecbd6b474ae7352f9",
                "responses": {
                    "200": {
                        "description": "Lista de cócteles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Mojito"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Cóctel clásico cubano"
                                            },
                                            "ingredients": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Ron Blanco"
                                                        },
                                                        "pivot": {
                                                            "properties": {
                                                                "measure_ml": {
                                                                    "type": "number",
                                                                    "example": 50
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cocktails"
                ],
                "summary": "Crear un nuevo cóctel",
                "description": "Permite a los administradores crear un cóctel y asociar sus ingredientes con cantidades en mililitros.",
                "operationId": "c63b237aa16074c5773b5b64034e8329",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Negroni"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Cóctel italiano amargo"
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "measure_ml": {
                                                    "type": "number",
                                                    "example": 30
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cóctel creado correctamente"
                    },
                    "403": {
                        "description": "Prohibido (solo admin)"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/cocktails/{id}": {
            "get": {
                "tags": [
                    "Cocktails"
                ],
                "summary": "Ver un cóctel específico",
                "description": "Muestra la información detallada de un cóctel con sus ingredientes.",
                "operationId": "3cb09df5796da56d8650ac2e07f1a6a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cóctel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cóctel encontrado correctamente"
                    },
                    "404": {
                        "description": "Cóctel no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cocktails"
                ],
                "summary": "Actualizar un cóctel existente",
                "description": "Permite a los administradores actualizar los datos de un cóctel y sus ingredientes asociados.",
                "operationId": "a99a8b597d4640f81a4376c5604a900c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cóctel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Negroni Actualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Versión mejorada del clásico Negroni"
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "measure_ml": {
                                                    "type": "number",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cóctel actualizado correctamente"
                    },
                    "404": {
                        "description": "Cóctel no encontrado"
                    },
                    "403": {
                        "description": "Prohibido (solo admin)"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cocktails"
                ],
                "summary": "Eliminar un cóctel",
                "description": "Permite a los administradores eliminar un cóctel del sistema.",
                "operationId": "dd5c8b76599b92946d6a69c32b47477a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cóctel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cóctel eliminado correctamente"
                    },
                    "404": {
                        "description": "Cóctel no encontrado"
                    },
                    "403": {
                        "description": "Prohibido (solo admin)"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/search/{ingredientId}": {
            "get": {
                "tags": [
                    "Cocktails"
                ],
                "summary": "Buscar cócteles por ingrediente",
                "description": "Permite buscar cócteles que contengan un ingrediente específico.",
                "operationId": "c0a6e619ea4e841665a2c0448c16a694",
                "parameters": [
                    {
                        "name": "ingredientId",
                        "in": "path",
                        "description": "ID del ingrediente a buscar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de cócteles que contienen el ingrediente"
                    },
                    "404": {
                        "description": "No se encontraron cócteles con ese ingrediente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/cocktails/{id}/alcohol-content": {
            "get": {
                "tags": [
                    "Cocktails"
                ],
                "summary": "Calcular contenido alcohólico de un cóctel",
                "description": "Calcula el contenido de alcohol en base al porcentaje de alcohol de sus ingredientes y sus cantidades (measure_ml).",
                "operationId": "56d3cfedf9a71fb20d54a0805e3519f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cóctel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Porcentaje alcohólico calculado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cocktail": {
                                            "type": "string",
                                            "example": "Margarita"
                                        },
                                        "alcohol_content": {
                                            "type": "string",
                                            "example": "12.5%"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cóctel no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "Usa el token Bearer devuelto tras iniciar sesión o registrarte.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "passport": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}